dom/tests/mochitest/cowl/test_labeled_data_response.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper(result, desc) {
    ok(result, desc);

    completedTests++;

    if (result) {
      passedTests++;
    }

    if (completedTests == 1) {
      is(passedTests, completedTests, "There are " + completedTests + " privilege tests that should pass");

      SimpleTest.finish();
    }
  }

fetch('http://example.com/tests/dom/tests/mochitest/cowl/labeled.html').then(function(res) {
  ok_wrapper(false, "Should not be able to load labeled response when not confined");
}).catch(function(e) {
  ok_wrapper(true, "Should not be able to load labeled response when not confined");
});
  </script>
</pre>
</body>
</html>
